+++ packages/orq-rc/RELEASES.md
@@ -6259,3 +6259,13 @@ Based on:
 - [python v4.5.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.21 - packages/orq-rc
+
+## 2026-03-03 08:18:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.739.0 (2.846.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.22 - packages/orq-rc